Small fix for corrupted areas
I don't know if this human heart is anywhere else in the game but the game use this when loads loads corrupted areas. With this small fix, there will be no more "corrupted areas" where they aren't. Also fixes Tora Master missions with the Den, announced like a corrupted zone.